add refactor from past (#113)
* add refactor from past

* refactor resign & reshare

* fix critics
