+++ pkgs/operator/state.go
@@ -2,7 +2,6 @@ package operator
 
 import (
 	"bytes"
-	"crypto/rand"
 	"crypto/rsa"
 	"encoding/hex"
 	"encoding/json"
@@ -12,10 +11,8 @@ import (
 
 	cli_utils "github.com/bloxapp/ssv-dkg/cli/utils"
 	"github.com/bloxapp/ssv-dkg/pkgs/crypto"
-	"github.com/bloxapp/ssv-dkg/pkgs/dkg"
 	"github.com/bloxapp/ssv-dkg/pkgs/utils"
 	"github.com/bloxapp/ssv-dkg/pkgs/wire"
-	"github.com/bloxapp/ssv/utils/rsaencryption"
 	"github.com/drand/kyber"
 	kyber_bls12381 "github.com/drand/kyber-bls12381"
 	kyber_dkg "github.com/drand/kyber/share/dkg"
@@ -29,47 +26,6 @@ import (
 
 const MaxInstances = 1024
 const MaxInstanceTime = 5 * time.Minute
-
-// Instance interface to process messages at DKG instances incoming from initiator
-type Instance interface {
-	Process(*wire.SignedTransport, []*spec.Operator) error
-	ReadResponse() []byte
-	ReadError() error
-	VerifyInitiatorMessage(msg, sig []byte) error
-	GetLocalOwner() *dkg.LocalOwner
-	CheckIncomingOperators([]*wire.SignedTransport) (map[uint64]*spec.Operator, error)
-}
-
-// instWrapper wraps LocalOwner instance with RSA public key
-type instWrapper struct {
-	*dkg.LocalOwner                   // main DKG ceremony instance
-	InitiatorPublicKey *rsa.PublicKey // initiator's RSA public key to verify its identity. Makes sure that in the DKG process messages received only from one initiator who started it.
-	respChan           chan []byte    // channel to receive response
-	errChan            chan error     // channel to receive error
-}
-
-// VerifyInitiatorMessage verifies initiator message signature
-func (iw *instWrapper) VerifyInitiatorMessage(msg, sig []byte) error {
-	pubKey, err := spec_crypto.EncodeRSAPublicKey(iw.InitiatorPublicKey)
-	if err != nil {
-		return err
-	}
-	if err := spec_crypto.VerifyRSA(iw.InitiatorPublicKey, msg, sig); err != nil {
-		return fmt.Errorf("failed to verify a message from initiator: %x", pubKey)
-	}
-	iw.Logger.Info("Successfully verified initiator message signature", zap.Uint64("from", iw.ID))
-	return nil
-}
-
-// ReadResponse reads from response channel
-func (iw *instWrapper) ReadResponse() []byte {
-	return <-iw.respChan
-}
-
-// ReadError reads from error channel
-func (iw *instWrapper) ReadError() error {
-	return <-iw.errChan
-}
 
 // InstanceID each new DKG ceremony has a unique random ID that we can identify messages and be able to process them in parallel
 type InstanceID [24]byte
@@ -85,68 +41,6 @@ type Switch struct {
 	PubKeyBytes      []byte
 	OperatorID       uint64
 	EthClient        eip1271.ETHClient
-}
-
-// CreateInstance creates a LocalOwner instance with the DKG ceremony ID, that we can identify it later. Initiator public key identifies an initiator for
-// new instance. There cant be two instances with the same ID, but one initiator can start several DKG ceremonies.
-func (s *Switch) CreateInstance(reqID [24]byte, operators []*spec.Operator, message interface{}, initiatorPublicKey *rsa.PublicKey) (Instance, []byte, error) {
-	operatorID, err := spec.OperatorIDByPubKey(operators, s.PubKeyBytes)
-	if err != nil {
-		return nil, nil, err
-	}
-	// sanity check of operator ID
-	if s.OperatorID != operatorID {
-		return nil, nil, fmt.Errorf("wrong operator ID: want %d, got %d", s.OperatorID, operatorID)
-	}
-	bchan := make(chan []byte, 1)
-	broadcast := func(msg []byte) error {
-		bchan <- msg
-		return nil
-	}
-	opts := dkg.OwnerOpts{
-		Logger:             s.Logger.With(zap.String("instance", hex.EncodeToString(reqID[:]))),
-		BroadcastF:         broadcast,
-		Signer:             crypto.RSASigner(s.PrivateKey),
-		EncryptFunc:        s.Encrypt,
-		DecryptFunc:        s.Decrypt,
-		Suite:              kyber_bls12381.NewBLS12381Suite(),
-		ID:                 operatorID,
-		InitiatorPublicKey: initiatorPublicKey,
-		OperatorSecretKey:  s.PrivateKey,
-		Version:            s.Version,
-	}
-	owner := dkg.New(&opts)
-	var resp *wire.Transport
-	// wait for exchange msg
-	switch msg := message.(type) {
-	case *spec.Init:
-		resp, err = owner.Init(reqID, msg)
-		if err != nil {
-			return nil, nil, err
-		}
-	case *wire.ResignMessage:
-		resp, err = owner.Resign(reqID, msg)
-		if err != nil {
-			return nil, nil, err
-		}
-	case *wire.ReshareMessage:
-		commits, share, err := s.getPublicCommitsAndSecretShare(msg)
-		if err != nil {
-			return nil, nil, err
-		}
-		owner.SecretShare = share
-		resp, err = owner.Reshare(reqID, &msg.SignedReshare.Reshare, commits)
-		if err != nil {
-			return nil, nil, err
-		}
-	default:
-		return nil, nil, fmt.Errorf("cant determine the ceremony message type")
-	}
-	if err := owner.Broadcast(resp); err != nil {
-		return nil, nil, err
-	}
-	res := <-bchan
-	return &instWrapper{owner, initiatorPublicKey, bchan, owner.ErrorChan}, res, nil
 }
 
 func (s *Switch) getPublicCommitsAndSecretShare(reshareMsg *wire.ReshareMessage) ([]kyber.Point, *kyber_dkg.DistKeyShare, error) {
@@ -201,21 +95,6 @@ func (s *Switch) getPublicCommitsAndSecretShare(reshareMsg *wire.ReshareMessage)
 	return commits, distKeyShare, err
 }
 
-// Sign creates a RSA signature for the message at operator before sending it to initiator
-func (s *Switch) Sign(msg []byte) ([]byte, error) {
-	return spec_crypto.SignRSA(s.PrivateKey, msg)
-}
-
-// Encrypt with RSA public key private DKG share key
-func (s *Switch) Encrypt(msg []byte) ([]byte, error) {
-	return rsa.EncryptPKCS1v15(rand.Reader, &s.PrivateKey.PublicKey, msg)
-}
-
-// Decrypt with RSA private key private DKG share key
-func (s *Switch) Decrypt(ciphertext []byte) ([]byte, error) {
-	return rsaencryption.DecodeKey(s.PrivateKey, ciphertext)
-}
-
 // NewSwitch creates a new Switch
 func NewSwitch(pv *rsa.PrivateKey, logger *zap.Logger, ver, pkBytes []byte, id uint64, ethClient eip1271.ETHClient) *Switch {
 	return &Switch{
@@ -229,60 +108,6 @@ func NewSwitch(pv *rsa.PrivateKey, logger *zap.Logger, ver, pkBytes []byte, id u
 		OperatorID:       id,
 		EthClient:        ethClient,
 	}
-}
-
-// InitInstance creates a LocalOwner instance and DKG public key message (Exchange)
-func (s *Switch) InitInstance(reqID [24]byte, initMsg *wire.Transport, initiatorPub, initiatorSignature []byte) ([]byte, error) {
-	if !bytes.Equal(initMsg.Version, s.Version) {
-		return nil, fmt.Errorf("wrong version: remote %s local %s", initMsg.Version, s.Version)
-	}
-	s.Logger.Info("🚀 Initializing Init instance")
-	init := &spec.Init{}
-	if err := init.UnmarshalSSZ(initMsg.Data); err != nil {
-		return nil, fmt.Errorf("init: failed to unmarshal init message: %s", err.Error())
-	}
-	if err := spec.ValidateInitMessage(init); err != nil {
-		return nil, err
-	}
-	// Check that incoming message signature is valid
-	initiatorPubKey, err := spec_crypto.ParseRSAPublicKey(initiatorPub)
-	if err != nil {
-		return nil, fmt.Errorf("init: failed parse initiator public key: %s", err.Error())
-	}
-	marshalledWireMsg, err := initMsg.MarshalSSZ()
-	if err != nil {
-		return nil, fmt.Errorf("init: failed to marshal transport message: %s", err.Error())
-	}
-	err = spec_crypto.VerifyRSA(initiatorPubKey, marshalledWireMsg, initiatorSignature)
-	if err != nil {
-		return nil, fmt.Errorf("init: initiator signature isn't valid: %s", err.Error())
-	}
-	s.Logger.Info("✅ init message signature is successfully verified", zap.String("from initiator", fmt.Sprintf("%x", initiatorPubKey.N.Bytes())))
-	if err := s.validateInstances(reqID); err != nil {
-		return nil, err
-	}
-	inst, resp, err := s.CreateInstance(reqID, init.Operators, init, initiatorPubKey)
-	if err != nil {
-		return nil, fmt.Errorf("init: failed to create instance: %s", err.Error())
-	}
-	s.Mtx.Lock()
-	s.Instances[reqID] = inst
-	s.InstanceInitTime[reqID] = time.Now()
-	s.Mtx.Unlock()
-	return resp, nil
-}
-
-// CleanInstances removes all instances at Switch
-func (s *Switch) CleanInstances() int {
-	count := 0
-	for id, instime := range s.InstanceInitTime {
-		if time.Now().After(instime.Add(MaxInstanceTime)) {
-			delete(s.Instances, id)
-			delete(s.InstanceInitTime, id)
-			count++
-		}
-	}
-	return count
 }
 
 // ProcessMessage processes incoming message to /dkg route
@@ -303,41 +128,7 @@ func (s *Switch) ProcessMessage(dkgMsg []byte) ([]byte, error) {
 	if !ok {
 		return nil, utils.ErrMissingInstance
 	}
-	var mltplMsgsBytes []byte
-	for _, ts := range st.Messages {
-		tsBytes, err := ts.MarshalSSZ()
-		if err != nil {
-			return nil, fmt.Errorf("process message: failed to marshal message: %s", err.Error())
-		}
-		mltplMsgsBytes = append(mltplMsgsBytes, tsBytes...)
-	}
-	// Verify initiator signature
-	err = inst.VerifyInitiatorMessage(mltplMsgsBytes, st.Signature)
-	if err != nil {
-		return nil, fmt.Errorf("process message: failed to verify initiator signature: %s", err.Error())
-	}
-	// check that we received enough messages from other operator participants
-	ops, err := inst.CheckIncomingOperators(st.Messages)
-	if err != nil {
-		return nil, err
-	}
-	var incOperators []*spec.Operator
-	for _, op := range ops {
-		incOperators = append(incOperators, op)
-	}
-	incOperators = spec.OrderOperators(incOperators)
-	if !spec.UniqueAndOrderedOperators(incOperators) {
-		return nil, fmt.Errorf("operators at incoming messages are not unique")
-	}
-	for _, ts := range st.Messages {
-		err = inst.Process(ts, incOperators)
-		if err != nil {
-			return nil, fmt.Errorf("process message: failed to process dkg message: %s", err.Error())
-		}
-	}
-	resp := inst.ReadResponse()
-
-	return resp, nil
+	return inst.ProcessMessages(st)
 }
 
 func (s *Switch) MarshallAndSign(msg wire.SSZMarshaller, msgType wire.TransportType, operatorID uint64, id [24]byte) ([]byte, error) {
@@ -463,159 +254,4 @@ func (s *Switch) VerifyIncomingMessage(incMsg *wire.SignedTransport) (uint64, er
 		return 0, err
 	}
 	return operatorID, nil
-}
-
-// InitInstance creates a LocalOwner instance and DKG public key message (Exchange)
-func (s *Switch) ResignInstance(reqID [24]byte, resignMsg *wire.Transport, initiatorPub, initiatorSignature []byte) ([]byte, error) {
-	if !bytes.Equal(resignMsg.Version, s.Version) {
-		return nil, fmt.Errorf("wrong version: remote %s local %s", resignMsg.Version, s.Version)
-	}
-	s.Logger.Info("🚀 Initializing Resigning instance")
-	// Check that incoming message signature is valid
-	initiatorPubKey, err := spec_crypto.ParseRSAPublicKey(initiatorPub)
-	if err != nil {
-		return nil, fmt.Errorf("resign: failed parse initiator public key: %s", err.Error())
-	}
-	marshalledWireMsg, err := resignMsg.MarshalSSZ()
-	if err != nil {
-		return nil, fmt.Errorf("resign: failed to marshal transport message: %s", err.Error())
-	}
-	err = spec_crypto.VerifyRSA(initiatorPubKey, marshalledWireMsg, initiatorSignature)
-	if err != nil {
-		return nil, fmt.Errorf("resign: initiator signature isn't valid: %s", err.Error())
-	}
-	resign := &wire.ResignMessage{}
-	if err := resign.UnmarshalSSZ(resignMsg.Data); err != nil {
-		return nil, fmt.Errorf("resign: failed to unmarshal init message: %s", err.Error())
-	}
-	s.Logger.Info("✅ resign message signature is successfully verified", zap.String("from initiator", fmt.Sprintf("%x", initiatorPubKey.N.Bytes())))
-	if err := s.validateInstances(reqID); err != nil {
-		return nil, err
-	}
-	s.Logger.Info("Incoming resign request fields",
-		zap.String("network", hex.EncodeToString(resign.SignedResign.Resign.Fork[:])),
-		zap.String("withdrawal", hex.EncodeToString(resign.SignedResign.Resign.WithdrawalCredentials)),
-		zap.String("owner", hex.EncodeToString(resign.SignedResign.Resign.Owner[:])),
-		zap.Uint64("nonce", resign.SignedResign.Resign.Nonce))
-	for _, proof := range resign.Proofs {
-		s.Logger.Info("Loaded proof",
-			zap.String("ValidatorPubKey", hex.EncodeToString(proof.Proof.ValidatorPubKey)),
-			zap.String("Owner", hex.EncodeToString(proof.Proof.Owner[:])),
-			zap.String("SharePubKey", hex.EncodeToString(proof.Proof.SharePubKey)),
-			zap.String("EncryptedShare", hex.EncodeToString(proof.Proof.EncryptedShare)),
-			zap.String("Signature", hex.EncodeToString(proof.Signature)))
-	}
-	// verify EIP1271 signature
-	if err := spec_crypto.VerifySignedMessageByOwner(
-		s.EthClient,
-		resign.SignedResign.Resign.Owner,
-		&resign.SignedResign.Resign,
-		resign.SignedResign.Signature,
-	); err != nil {
-		return nil, err
-	}
-	s.Logger.Info("✅ reshare eip1271 owner signature is successfully verified", zap.String("from initiator", fmt.Sprintf("%x", initiatorPubKey.N.Bytes())))
-	if err := s.validateInstances(reqID); err != nil {
-		return nil, err
-	}
-	inst, resp, err := s.CreateInstance(reqID, resign.Operators, resign, initiatorPubKey)
-	if err != nil {
-		return nil, fmt.Errorf("resign: failed to create instance: %s", err.Error())
-	}
-	s.Mtx.Lock()
-	s.Instances[reqID] = inst
-	s.InstanceInitTime[reqID] = time.Now()
-	s.Mtx.Unlock()
-	return resp, nil
-}
-
-// InitInstance creates a LocalOwner instance and DKG public key message (Exchange)
-func (s *Switch) ReshareInstance(reqID [24]byte, reshareMsg *wire.Transport, initiatorPub, initiatorSignature []byte) ([]byte, error) {
-	if !bytes.Equal(reshareMsg.Version, s.Version) {
-		return nil, fmt.Errorf("wrong version: remote %s local %s", reshareMsg.Version, s.Version)
-	}
-	// Check that incoming message signature is valid
-	initiatorPubKey, err := spec_crypto.ParseRSAPublicKey(initiatorPub)
-	if err != nil {
-		return nil, fmt.Errorf("resign: failed parse initiator public key: %s", err.Error())
-	}
-	marshalledWireMsg, err := reshareMsg.MarshalSSZ()
-	if err != nil {
-		return nil, fmt.Errorf("resign: failed to marshal transport message: %s", err.Error())
-	}
-	err = spec_crypto.VerifyRSA(initiatorPubKey, marshalledWireMsg, initiatorSignature)
-	if err != nil {
-		return nil, fmt.Errorf("resign: initiator signature isn't valid: %s", err.Error())
-	}
-	s.Logger.Info("🚀 Creating reshare instance")
-	reshare := &wire.ReshareMessage{}
-	if err := reshare.UnmarshalSSZ(reshareMsg.Data); err != nil {
-		return nil, fmt.Errorf("init: failed to unmarshal init message: %s", err.Error())
-	}
-	s.Logger.Info("Incoming reshare request fields",
-		zap.Any("Old operator IDs", utils.GetOpIDs(reshare.SignedReshare.Reshare.OldOperators)),
-		zap.Any("New operator IDs", utils.GetOpIDs(reshare.SignedReshare.Reshare.NewOperators)),
-		zap.String("ValidatorPubKey", hex.EncodeToString(reshare.Proofs[0].Proof.ValidatorPubKey)),
-		zap.String("network", hex.EncodeToString(reshare.SignedReshare.Reshare.Fork[:])),
-		zap.String("withdrawal", hex.EncodeToString(reshare.SignedReshare.Reshare.WithdrawalCredentials)),
-		zap.String("owner", hex.EncodeToString(reshare.SignedReshare.Reshare.Owner[:])),
-		zap.Uint64("nonce", reshare.SignedReshare.Reshare.Nonce),
-		zap.String("EIP1271 owner signature", hex.EncodeToString(reshare.SignedReshare.Signature)))
-	for _, proof := range reshare.Proofs {
-		s.Logger.Info("Reshare proof",
-			zap.String("ValidatorPubKey", hex.EncodeToString(proof.Proof.ValidatorPubKey)),
-			zap.String("Owner", hex.EncodeToString(proof.Proof.Owner[:])),
-			zap.String("SharePubKey", hex.EncodeToString(proof.Proof.SharePubKey)),
-			zap.String("EncryptedShare", hex.EncodeToString(proof.Proof.EncryptedShare)),
-			zap.String("Signature", hex.EncodeToString(proof.Signature)))
-	}
-	// verify EIP1271 signature
-	if err := spec_crypto.VerifySignedMessageByOwner(
-		s.EthClient,
-		reshare.SignedReshare.Reshare.Owner,
-		&reshare.SignedReshare.Reshare,
-		reshare.SignedReshare.Signature,
-	); err != nil {
-		return nil, err
-	}
-	s.Logger.Info("✅ reshare eip1271 owner signature is successfully verified", zap.String("from initiator", fmt.Sprintf("%x", initiatorPubKey.N.Bytes())))
-	if err := s.validateInstances(reqID); err != nil {
-		return nil, err
-	}
-	var allOps []*spec.Operator
-	allOps = append(allOps, reshare.SignedReshare.Reshare.OldOperators...)
-	allOps = append(allOps, reshare.SignedReshare.Reshare.NewOperators...)
-	inst, resp, err := s.CreateInstance(reqID, allOps, reshare, initiatorPubKey)
-	if err != nil {
-		return nil, fmt.Errorf("reshare: failed to create instance: %s", err.Error())
-	}
-	s.Mtx.Lock()
-	s.Instances[reqID] = inst
-	s.InstanceInitTime[reqID] = time.Now()
-	s.Mtx.Unlock()
-	return resp, nil
-}
-
-func (s *Switch) validateInstances(reqID InstanceID) error {
-	s.Mtx.Lock()
-	l := len(s.Instances)
-	if l >= MaxInstances {
-		cleaned := s.CleanInstances()
-		if l-cleaned >= MaxInstances {
-			s.Mtx.Unlock()
-			return utils.ErrMaxInstances
-		}
-	}
-	_, ok := s.Instances[reqID]
-	if ok {
-		tm := s.InstanceInitTime[reqID]
-		if time.Now().Before(tm.Add(MaxInstanceTime)) {
-			s.Mtx.Unlock()
-			return utils.ErrAlreadyExists
-		}
-		delete(s.Instances, reqID)
-		delete(s.InstanceInitTime, reqID)
-	}
-	s.Mtx.Unlock()
-	return nil
 }
