+++ pkgs/operator/state_test.go
@@ -229,12 +229,12 @@ func TestSwitch_cleanInstances(t *testing.T) {
 	resp, err := swtch.InitInstance(reqID, initMessage, encPubKey, sig)
 	require.NoError(t, err)
 	require.NotNil(t, resp)
-	require.Equal(t, swtch.CleanInstances(), 0)
+	require.Equal(t, swtch.cleanInstances(), 0)
 
 	require.Len(t, swtch.Instances, 1)
 	swtch.InstanceInitTime[reqID] = time.Now().Add(-time.Minute * 6)
 
-	require.Equal(t, swtch.CleanInstances(), 1)
+	require.Equal(t, swtch.cleanInstances(), 1)
 	require.Len(t, swtch.Instances, 0)
 
 }
